cf961d.py 几何题 枚举+判断三点共线

+++ tools/gen_code_tool/BaseGenTemplate.py
@@ -49,9 +49,9 @@ use std::io::{{BufRead, BufWriter, Write}};
 #[allow(unused)]
 pub fn solve(scan: &mut Scanner<impl BufRead>, out: &mut impl Write) {{
     let n = scan.token::<usize>();
-    let mut a = vec![0i32; n];
+    let mut a = vec![0i64; n];
     for i in 0..n {{
-        a[i] = scan.token::<i32>();
+        a[i] = scan.token::<i64>();
     }}   
     let mut ans = 0;
     writeln!(out, "{{}}", ans).ok();
